requirements.txt
torch>=1.13.0
torchvision>=0.14.0
streamlit>=1.18.0
scikit-learn>=1.0.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.60.0
